requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
mysqlclient==2.2.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Async & Caching
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-celery-beat==2.5.0

# Server
gunicorn==21.2.0

# API Documentation
drf-yasg==1.21.7

# Environment
python-decouple==3.8

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Monitoring
django-extensions==3.2.3

Pillow>=10.0.0
